pyproject.toml
[tool.poetry]
name = "api"
version = "0.1.0"
description = ""
authors = ["Nate Brooks", "Logan Parker"]

[tool.poetry.dependencies]
python = "^3.11"
django = "4.1.12"
django-cors-headers = "^4.2.0"
psycopg2 = "^2.9.9"
django-cleanup = "^8.0.0"
colorama = "^0.4.6"
channels = "^4.0.0"
djangorestframework = "^3.14.0"
pillow = "^10.0.1"
django-admin-inline-paginator = "^0.4.0"
django-storages = "^1.14.1"
boto3 = "^1.28.62"
firebase-admin = "^6.2.0"
django-filter = "^23.3"
django-anymail = {extras = ["mailgun"], version = "^10.1"}
django-extensions = "^3.2.3"
pydot = "^1.4.2"
redis = "^5.0.1"
google-api-python-client = "^2.102.0"
google-auth-httplib2 = "^0.1.1"
google-auth-oauthlib = "^1.1.0"
sendbird-platform-sdk = "^0.0.16"
python-magic = "^0.4.27"
freezegun = "^1.2.2"
wand = "^0.6.11"
ipython = "^8.16.1"
pyotp = "^2.9.0"
qrcode = "^7.4.2"
twilio = "^8.9.1"
authlib = "^1.2.1"
daphne = "^4.0.0"


[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
